\name{AnyTimeSeries-class}
\Rdversion{1.1}
\docType{class}
\alias{AnyTimeSeries-class}

\title{Class AnyTimeSeries}
\description{Class AnyTimeSeries.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "AnyTimeSeries" in the signature.
}
\author{Georgi N. Boshnakov}
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
\examples{
showClass("AnyTimeSeries")
}
\keyword{classes}
\keyword{internal}
