\name{as_date-methods}
\docType{methods}
\alias{as_date-methods}
\alias{as_date,PeriodicTimeSeries-method}
\title{Replace methods for as_date in package pcts}
\description{Replace methods for as_date in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicTimeSeries")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
