\name{as_datetime-methods}
\docType{methods}
\alias{as_datetime-methods}
\alias{as_datetime,PeriodicTimeSeries-method}
\title{Methods for as_datetime in package pcts}
\description{Methods for as_datetime in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicTimeSeries")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{chron}
