\name{backwardPartialCoefficients-methods}
\docType{methods}
\alias{backwardPartialCoefficients-methods}
\alias{backwardPartialCoefficients,VirtualPeriodicAutocovariances-method}
\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Methods}{
\describe{

\item{\code{signature(x = "VirtualPeriodicAutocovariances")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{internal}
