\name{date<--methods}
\docType{methods}
\alias{date<--methods}
\alias{date<-,BasicCycle-method}
\alias{date<-,Cyclic-method}
\title{Replace methods for date in package pcts}
\description{Replace methods for date in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "BasicCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "Cyclic")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
