\name{filterPolyCoef-methods}
\docType{methods}
\alias{filterPolyCoef-methods}
\alias{filterPolyCoef,PeriodicBJFilter-method}
\alias{filterPolyCoef,PeriodicSPFilter-method}
\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Methods}{
\describe{

\item{\code{signature(object = "PeriodicBJFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "PeriodicSPFilter")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{internal}
