\name{head-methods}
\docType{methods}
\alias{head-methods}
\alias{head,PeriodicTimeSeries-method}
\title{Methods for function head() in package pcts}
\description{Methods for function head() in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicTimeSeries")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{internal}
