\name{LegacyPeriodicFilterModel-class}
\Rdversion{1.1}
\docType{class}
\alias{LegacyPeriodicFilterModel-class}

\title{Class LegacyPeriodicFilterModel}
\description{Class LegacyPeriodicFilterModel.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("LegacyPeriodicFilterModel", ...)}.
%%  ~~ describe objects here ~~
}
\section{Slots}{
  \describe{
    \item{\code{intercept}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{arFilter}:}{Object of class \code{"PeriodicArFilter"} ~~ }
    \item{\code{maFilter}:}{Object of class \code{"PeriodicMaFilter"} ~~ }
    \item{\code{scalesq}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{jordan}:}{Object of class \code{"ANY"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{VirtualPeriodicFilterModel}"}, directly.
}
\section{Methods}{
No methods defined with class "LegacyPeriodicFilterModel" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("LegacyPeriodicFilterModel")
%}
\keyword{classes}
\keyword{internal}
