\name{maxLag-methods}
\docType{methods}
\alias{maxLag-methods}
\alias{maxLag,PeriodicArmaFilter-method}


\title{Methods for function maxLag() in package 'pcts'}
\description{Methods for function maxLag() in package 'pcts'.}
\section{Methods}{
\describe{

\item{\code{signature(object = "PeriodicArmaFilter")}}{{
%%  ~~describe this method here~~
}
}}}
\keyword{methods}
