\name{ModelCycleSpec-class}
\Rdversion{1.1}
\docType{class}
\alias{ModelCycleSpec-class}
%\alias{modelCycle,ModelCycleSpec-method}
%\alias{modelCycle<-,ModelCycleSpec-method}

\title{Class ModelCycleSpec}
\description{Class ModelCycleSpec.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ModelCycleSpec", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{modelCycle}:}{Object of class \code{"BasicCycle"} ~~ }
  }
}
\section{Methods}{
  \describe{
    \item{modelCycle}{\code{signature(object = "ModelCycleSpec")}: ... }
    \item{modelCycle<-}{\code{signature(object = "ModelCycleSpec")}: ... }
	 }
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("ModelCycleSpec")
%}
\keyword{classes}
