\name{nTicks-methods}
\docType{methods}
\alias{nTicks-methods}
\alias{nTicks,matrix-method}
\alias{nTicks,numeric-method}
\alias{nTicks,PeriodicTimeSeries-method}
\alias{nTicks,Cyclic-method}
\title{Number of observations in a time series}
\description{Number of observations in a time series.}
\section{Methods}{
\describe{

\item{\code{signature(x = "matrix")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "numeric")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "PeriodicTimeSeries")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(x = "Cyclic")}}{{
%%  ~~describe this method here~~
}
}}}
\keyword{methods}
\keyword{internal}
