\name{partialAutocorrelations-methods}
\docType{methods}
\alias{partialAutocorrelations-methods}
\alias{partialAutocorrelations,PeriodicAutocovariances,ANY,missing-method}
\alias{partialAutocorrelations,VirtualPeriodicAutocovariances,ANY,ANY-method}
\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicAutocovariances", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "VirtualPeriodicAutocovariances", maxlag = "ANY", lag_0 = "ANY")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{internal}
