\name{partialAutocovariances-methods}
\docType{methods}
\alias{partialAutocovariances-methods}
\alias{partialAutocovariances,VirtualPeriodicAutocovariances-method}
\title{
    Compute periodic partial autocovariances
}
\description{
  Compute periodic partial autocovariances.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "VirtualPeriodicAutocovariances")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link[sarima]{partialAutocorrelations}} in package \pkg{sarima}
  for further details.
  
  \code{\link{partialAutocorrelations}},
  \code{\link{partialVariances}}
}
\keyword{methods}
