\name{partialCoefficients-methods}
\docType{methods}
\alias{partialCoefficients-methods}
\alias{partialCoefficients,PeriodicArModel-method}
\alias{partialCoefficients,VirtualPeriodicAutocovariances-method}
\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicArModel")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "VirtualPeriodicAutocovariances")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{internal}
