\name{PeriodicArmaFilter-class}
\Rdversion{1.1}
\docType{class}
\alias{PeriodicArmaFilter-class}
\alias{PeriodicArFilter-class}
\alias{PeriodicMaFilter-class}

\title{Class \code{"PeriodicArmaFilter"}}
\description{Class PeriodicArmaFilter.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("PeriodicArmaFilter", ..., ar, ma, nseasons)}.
%%  ~~ describe objects here ~~
}
\section{Slots}{
  \describe{
    \item{\code{ar}:}{Object of class \code{"PeriodicBJFilter"} ~~ }
    \item{\code{ma}:}{Object of class \code{"PeriodicSPFilter"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{VirtualArmaFilter}"}, directly.
Class \code{"\linkS4class{VirtualMonicFilter}"}, by class "VirtualArmaFilter", distance 2.
}
\section{Methods}{
No methods defined with class "PeriodicArmaFilter" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("PeriodicArmaFilter")
%}
\keyword{classes}
