\name{PeriodicAutocovarianceSpec-class}
\Rdversion{1.1}
\docType{class}
\alias{PeriodicAutocovarianceSpec-class}

\title{Class PeriodicAutocovarianceSpec}
\description{Class PeriodicAutocovarianceSpec.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("PeriodicAutocovarianceSpec", ...)}.
%%  ~~ describe objects here ~~
}
\section{Slots}{
  \describe{
    \item{\code{pcmatrix}:}{Object of class \code{"ANY"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "PeriodicAutocovarianceSpec" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("PeriodicAutocovarianceSpec")
%}
\keyword{classes}
