\name{PeriodicIntegratedArmaSpec-class}
\Rdversion{1.1}
\docType{class}
\alias{PeriodicIntegratedArmaSpec-class}
%\alias{nSeasons,PeriodicIntegratedArmaSpec-method}

\title{Class PeriodicIntegratedArmaSpec}
\description{Class PeriodicIntegratedArmaSpec.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("PeriodicIntegratedArmaSpec", ...)}.
%%  ~~ describe objects here ~~
}
\section{Slots}{
  \describe{
    \item{\code{pcmodel}:}{Object of class \code{"PeriodicArmaModel"} ~~ }
  }
}
\section{Methods}{
%%%   \describe{
%%%     \item{nSeasons}{\code{signature(object = "PeriodicIntegratedArmaSpec")}: ... }
%%% 	 }
\describe{
    \item{sigmaSq}{\code{signature(object = "PeriodicIntegratedArmaSpec")}: ... }}}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("PeriodicIntegratedArmaSpec")
%}
\keyword{classes}
