\name{PiPeriodicArmaModel-class}
\Rdversion{1.1}
\docType{class}
\alias{PiPeriodicArmaModel-class}

\title{Class PiPeriodicArmaModel}
\description{Class PiPeriodicArmaModel.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("PiPeriodicArmaModel", ...)}.
%%  ~~ describe objects here ~~
}
\section{Slots}{
  \describe{
    \item{\code{piorder}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{picoef}:}{Object of class \code{"matrix"} ~~ }
    \item{\code{pcmodel}:}{Object of class \code{"PeriodicArmaModel"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{VirtualPeriodicFilterModel}"}, directly.
Class \code{"\linkS4class{PeriodicIntegratedArmaSpec}"}, directly.
}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "PiPeriodicArmaModel")}: ... }
  }
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("PiPeriodicArmaModel")
%}
\keyword{classes}
