\name{SamplePeriodicAutocorrelations-class}
\Rdversion{1.1}
\docType{class}
\alias{SamplePeriodicAutocorrelations-class}

\title{Class SamplePeriodicAutocorrelations}
\description{Class SamplePeriodicAutocorrelations.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("SamplePeriodicAutocorrelations", ..., data)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{modelCycle}:}{Object of class \code{"BasicCycle"} ~~ }
    \item{\code{data}:}{Object of class \code{"Lagged"} ~~ }
    \item{\code{n}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{varnames}:}{Object of class \code{"character"} ~~ }
    \item{\code{objectname}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{PeriodicAutocorrelations}"}, directly.
Class \code{"\linkS4class{Fitted}"}, directly.
Class \code{"\linkS4class{ModelCycleSpec}"}, by class "PeriodicAutocorrelations", distance 2.
Class \code{"\linkS4class{FlexibleLagged}"}, by class "PeriodicAutocorrelations", distance 2.
Class \code{"\linkS4class{VirtualPeriodicAutocorrelations}"}, by class "PeriodicAutocorrelations", distance 2.
Class \code{"\linkS4class{Lagged}"}, by class "PeriodicAutocorrelations", distance 3.
Class \code{"\linkS4class{VirtualPeriodicModel}"}, by class "PeriodicAutocorrelations", distance 3.
}
\section{Methods}{
No methods defined with class "SamplePeriodicAutocorrelations" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("SamplePeriodicAutocorrelations")
%}
\keyword{classes}
