\name{seqSeasons-methods}
\docType{methods}
\alias{seqSeasons-methods}
\alias{seqSeasons,BasicCycle-method}
\alias{seqSeasons,Cyclic-method}
\alias{seqSeasons,VirtualPeriodicModel-method}
\title{Methods for seqSeasons() in package pcts}
\description{Methods for seqSeasons() in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "BasicCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "Cyclic")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "VirtualPeriodicModel")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{allSeasons}} for related functions and examples
}
\keyword{methods}
