\name{sigmaSq-methods}
\docType{methods}
\alias{sigmaSq-methods}
\alias{sigmaSq,PeriodicIntegratedArmaSpec-method}
\alias{sigmaSq,PeriodicInterceptSpec-method}
\title{Methods for \code{sigmaSq} in package pcts}
\description{Methods for \code{sigmaSq} in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(object = "PeriodicIntegratedArmaSpec")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "PeriodicInterceptSpec")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
