\name{SiPeriodicArmaModel-class}
\Rdversion{1.1}
\docType{class}
\alias{SiPeriodicArmaModel-class}

\title{Class SiPeriodicArmaModel}
\description{Class SiPeriodicArmaModel.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("SiPeriodicArmaModel", ...)}.
%%  ~~ describe objects here ~~
}
\section{Slots}{
  \describe{
    \item{\code{iorder}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{siorder}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{pcmodel}:}{Object of class \code{"PeriodicArmaModel"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{VirtualPeriodicFilterModel}"}, directly.
Class \code{"\linkS4class{PeriodicIntegratedArmaSpec}"}, directly.
}
\section{Methods}{
No methods defined with class "SiPeriodicArmaModel" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("SiPeriodicArmaModel")
%}
\keyword{classes}
