\name{SLTypeMatrix-class}
\Rdversion{1.1}
\docType{class}
\alias{SLTypeMatrix-class}

\title{Class SLTypeMatrix}
\description{
  Class SLTypeMatrix is a virtual class for classes that can hold
  objects with season-lag conventions. 
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "SLTypeMatrix" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
\examples{
showClass("SLTypeMatrix")
}
\keyword{classes}
\keyword{internal}
