\name{tail-methods}
\docType{methods}
\alias{tail-methods}
\alias{tail,PeriodicTimeSeries-method}
\title{Methods for function tail() in package pcts}
\description{Methods for function tail() in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicTimeSeries")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{internal}
