\name{unitCycle-methods}
\docType{methods}
\alias{unitCycle-methods}
\alias{unitCycle,ANY-method}
\alias{unitCycle,DayWeekCycle-method}
\alias{unitCycle,FiveDayWeekCycle-method}
\alias{unitCycle,MonthYearCycle-method}
\alias{unitCycle,OpenCloseCycle-method}
\alias{unitCycle,QuarterYearCycle-method}
\alias{unitCycle,SimpleCycle-method}
\alias{unitCycle,Cyclic-method}
\alias{unitCycle,Every30MinutesCycle-method}
\alias{unitCycle,VirtualPeriodicModel-method}
\title{Methods for unitCycle() in package pcts}
\description{Methods for unitCycle() in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "DayWeekCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "FiveDayWeekCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "MonthYearCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "OpenCloseCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "QuarterYearCycle")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(x = "SimpleCycle")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "Cyclic")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "Every30MinutesCycle")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualPeriodicModel")}}{{
%%  ~~describe this method here~~
}
}}}
\seealso{
  \code{\link{allSeasons}} for related functions and examples.
}
\keyword{methods}
%\keyword{internal}
