\name{unitSeason-methods}
\docType{methods}
\alias{unitSeason-methods}
\alias{unitSeason,ANY-method}
\alias{unitSeason,DayWeekCycle-method}
\alias{unitSeason,FiveDayWeekCycle-method}
\alias{unitSeason,MonthYearCycle-method}
\alias{unitSeason,OpenCloseCycle-method}
\alias{unitSeason,QuarterYearCycle-method}
\alias{unitSeason,SimpleCycle-method}
\alias{unitSeason,Cyclic-method}
\alias{unitSeason,Every30MinutesCycle-method}
\alias{unitSeason,VirtualPeriodicModel-method}
\title{Methods for unitSeason() in package pcts}
\description{Methods for unitSeason() in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "DayWeekCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "FiveDayWeekCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "MonthYearCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "OpenCloseCycle")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "QuarterYearCycle")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(x = "SimpleCycle")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "Cyclic")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "Every30MinutesCycle")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualPeriodicModel")}}{{
%%  ~~describe this method here~~
}
}}}
\seealso{
  \code{\link{allSeasons}} for related functions and examples
}
\keyword{methods}
%\keyword{internal}
