\name{unitSeason<--methods}
\docType{methods}
\alias{unitSeason<--methods}
\alias{unitSeason<-,Cyclic-method}
\alias{unitSeason<-,SimpleCycle-method}

% 2020-02-16 use backticks below, since pkgdown doesn't render well
%     double quotes in this context.
\title{Methods for \code{`unitSeason<-`} in package pcts}
\description{Methods for \code{`unitSeason<-`} in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "Cyclic")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "SimpleCycle")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
