\name{VirtualPeriodicArmaModel-class}
\Rdversion{1.1}
\docType{class}
\alias{VirtualPeriodicArmaModel-class}
\alias{pcMean,VirtualPeriodicArmaModel-method}

\title{Class VirtualPeriodicArmaModel}
\description{Class VirtualPeriodicArmaModel.}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("VirtualPeriodicArmaModel", ...)}.
%%  ~~ describe objects here ~~
}
\section{Extends}{
Class \code{"\linkS4class{VirtualPeriodicFilterModel}"}, directly.
Class \code{"\linkS4class{VirtualPeriodicStationaryModel}"}, directly.
Class \code{"\linkS4class{VirtualPeriodicAutocovarianceModel}"}, by class "VirtualPeriodicStationaryModel", distance 2.
Class \code{"\linkS4class{VirtualPeriodicMeanModel}"}, by class "VirtualPeriodicStationaryModel", distance 2.
}
\section{Methods}{
  \describe{
    \item{PeriodicArModel}{\code{signature(object = "VirtualPeriodicArmaModel")}: ... }}
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("VirtualPeriodicArmaModel")
%}
\keyword{classes}
\keyword{internal}
