\name{VirtualPeriodicArModel-class}
\Rdversion{1.1}
\docType{class}
\alias{VirtualPeriodicArModel-class}

\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("VirtualPeriodicArModel", ...)}.
%%  ~~ describe objects here ~~
}
\section{Extends}{
Class \code{"\linkS4class{VirtualPeriodicArmaModel}"}, directly.
Class \code{"\linkS4class{VirtualPeriodicFilterModel}"}, by class "VirtualPeriodicArmaModel", distance 2.
Class \code{"\linkS4class{VirtualPeriodicStationaryModel}"}, by class "VirtualPeriodicArmaModel", distance 2.
Class \code{"\linkS4class{VirtualPeriodicAutocovarianceModel}"}, by class "VirtualPeriodicArmaModel", distance 3.
Class \code{"\linkS4class{VirtualPeriodicMeanModel}"}, by class "VirtualPeriodicArmaModel", distance 3.
}
\section{Methods}{
No methods defined with class "VirtualPeriodicArModel" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("VirtualPeriodicArModel")
%}
\keyword{classes}
\keyword{internal}
