\name{VirtualPeriodicAutocorrelations-class}
\Rdversion{1.1}
\docType{class}
\alias{VirtualPeriodicAutocorrelations-class}

\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{VirtualPeriodicModel}"}, directly.
}
\section{Methods}{
No methods defined with class "VirtualPeriodicAutocorrelations" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("VirtualPeriodicAutocorrelations")
%}
\keyword{classes}
\keyword{internal}
