\name{VirtualPeriodicAutocovarianceModel-class}
\Rdversion{1.1}
\docType{class}
\alias{VirtualPeriodicAutocovarianceModel-class}
%\alias{[,VirtualPeriodicAutocovarianceModel,missing,missing,ANY-method}
%\alias{[,VirtualPeriodicAutocovarianceModel,missing,numeric,ANY-method}
%\alias{[,VirtualPeriodicAutocovarianceModel,numeric,missing,ANY-method}
%\alias{[,VirtualPeriodicAutocovarianceModel,numeric,numeric,ANY-method}
%\alias{autocorrelations,VirtualPeriodicAutocovarianceModel-method}
%\alias{maxLag,VirtualPeriodicAutocovarianceModel-method}

\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{[}{\code{signature(x = "VirtualPeriodicAutocovarianceModel", i = "missing", j = "missing", drop = "ANY")}: ... }
    \item{[}{\code{signature(x = "VirtualPeriodicAutocovarianceModel", i = "missing", j = "numeric", drop = "ANY")}: ... }
    \item{[}{\code{signature(x = "VirtualPeriodicAutocovarianceModel", i = "numeric", j = "missing", drop = "ANY")}: ... }
    \item{[}{\code{signature(x = "VirtualPeriodicAutocovarianceModel", i = "numeric", j = "numeric", drop = "ANY")}: ... }
    \item{maxLag}{\code{signature(object = "VirtualPeriodicAutocovarianceModel")}: ... }
	 
    \item{autocorrelations}{\code{signature(x = "VirtualPeriodicAutocovarianceModel", maxlag = "missing", lag_0 = "ANY")}: ... }
    \item{autocorrelations}{\code{signature(x = "VirtualPeriodicAutocovarianceModel", maxlag = "numeric", lag_0 = "ANY")}: ... }}
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("VirtualPeriodicAutocovarianceModel")
%}
\keyword{classes}
\keyword{internal}
