\name{VirtualPeriodicAutocovariances-class}
\Rdversion{1.1}
\docType{class}
\alias{VirtualPeriodicAutocovariances-class}
%\alias{autocorrelations,VirtualPeriodicAutocovariances,ANY,missing-method}
%\alias{autocovariances,VirtualPeriodicAutocovariances,ANY-method}
%\alias{backwardPartialCoefficients,VirtualPeriodicAutocovariances-method}
%\alias{backwardPartialVariances,VirtualPeriodicAutocovariances-method}
%\alias{partialAutocorrelations,VirtualPeriodicAutocovariances,ANY,ANY-method}
%\alias{partialAutocovariances,VirtualPeriodicAutocovariances-method}
%\alias{partialCoefficients,VirtualPeriodicAutocovariances-method}
%\alias{partialVariances,VirtualPeriodicAutocovariances-method}
%\alias{pc.bU,VirtualPeriodicAutocovariances-method}
%\alias{pc.fL,VirtualPeriodicAutocovariances-method}
%\alias{pc.phis2,VirtualPeriodicAutocovariances-method}

\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{VirtualPeriodicModel}"}, directly.
}
\section{Methods}{
  \describe{
    \item{autocorrelations}{\code{signature(x = "VirtualPeriodicAutocovariances", maxlag = "ANY", lag_0 = "missing")}: ... }
    \item{autocovariances}{\code{signature(x = "VirtualPeriodicAutocovariances", maxlag = "ANY")}: ... }
    \item{backwardPartialCoefficients}{\code{signature(x = "VirtualPeriodicAutocovariances")}: ... }
    \item{backwardPartialVariances}{\code{signature(x = "VirtualPeriodicAutocovariances")}: ... }
    \item{partialAutocorrelations}{\code{signature(x = "VirtualPeriodicAutocovariances", maxlag = "ANY", lag_0 = "ANY")}: ... }
    \item{partialAutocovariances}{\code{signature(x = "VirtualPeriodicAutocovariances")}: ... }
    \item{partialCoefficients}{\code{signature(x = "VirtualPeriodicAutocovariances")}: ... }
    \item{partialVariances}{\code{signature(x = "VirtualPeriodicAutocovariances")}: ... }
    \item{pc.bU}{\code{signature(x = "VirtualPeriodicAutocovariances")}: ... }
    \item{pc.fL}{\code{signature(x = "VirtualPeriodicAutocovariances")}: ... }
    \item{pc.phis2}{\code{signature(x = "VirtualPeriodicAutocovariances")}: ... }
	 }
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("VirtualPeriodicAutocovariances")
%}
\keyword{classes}
\keyword{internal}
