\name{VirtualPeriodicModel-class}
\Rdversion{1.1}
\docType{class}
\alias{VirtualPeriodicModel-class}
%\alias{allSeasons,VirtualPeriodicModel,ANY-method}
%\alias{nSeasons,VirtualPeriodicModel-method}
%\alias{seqSeasons,VirtualPeriodicModel-method}
%\alias{unitCycle,VirtualPeriodicModel-method}
%\alias{unitSeason,VirtualPeriodicModel-method}

\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{allSeasons}{\code{signature(x = "VirtualPeriodicModel", abb = "ANY")}: ... }
    \item{nSeasons}{\code{signature(object = "VirtualPeriodicModel")}: ... }
    \item{seqSeasons}{\code{signature(x = "VirtualPeriodicModel")}: ... }
    \item{unitCycle}{\code{signature(x = "VirtualPeriodicModel")}: ... }
    \item{unitSeason}{\code{signature(x = "VirtualPeriodicModel")}: ... }
	 }
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("VirtualPeriodicModel")
%}
\keyword{classes}
\keyword{internal}
