\name{VirtualPeriodicWhiteNoiseModel-class}
\Rdversion{1.1}
\docType{class}
\alias{VirtualPeriodicWhiteNoiseModel-class}

\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("VirtualPeriodicWhiteNoiseModel", ...)}.
%%  ~~ describe objects here ~~
}
\section{Extends}{
Class \code{"\linkS4class{VirtualPeriodicStationaryModel}"}, directly.
Class \code{"\linkS4class{VirtualPeriodicAutocovarianceModel}"}, by class "VirtualPeriodicStationaryModel", distance 2.
Class \code{"\linkS4class{VirtualPeriodicMeanModel}"}, by class "VirtualPeriodicStationaryModel", distance 2.
}
\section{Methods}{
No methods defined with class "VirtualPeriodicWhiteNoiseModel" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("VirtualPeriodicWhiteNoiseModel")
%}
\keyword{classes}
\keyword{internal}
