\name{zoo-class}
\Rdversion{1.1}
\docType{class}
\alias{zoo-class}
\title{Class zoo made S4}
\description{Class zoo made S4}
\section{
  Objects from the Class}{A virtual Class: No objects may be created
  from it.

  S4 Class \code{"zoo"} is derived from its namesake in package
  \pkg{zoo} for use as a super class for periodic time series classes
  and in S4 method signatures.

}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
No methods defined with class "zoo" in the signature.
}
\seealso{
    \code{\linkS4class{PeriodicTS_zooreg}},
    \code{\linkS4class{zooreg}}
    and package \pkg{zoo}
}
\examples{
showClass("zoo")
}
\keyword{classes}
