\name{[-methods}
\docType{methods}
\alias{[-methods}
\alias{[,BasicCycle,ANY,missing,ANY-method}
\alias{[,BasicCycle,missing,missing,ANY-method}
\alias{[,PeriodicMTS,ANY,missing,ANY-method}
\alias{[,PeriodicMTS,missing,missing,ANY-method}
\alias{[,PeriodicVector,ANY,ANY,ANY-method}
\alias{[,PeriodicVector,ANY,missing,ANY-method}
\alias{[,PeriodicVector,missing,ANY,ANY-method}
\alias{[,PeriodicVector,missing,missing,ANY-method}
\alias{[,VirtualPeriodicAutocovarianceModel,missing,missing,ANY-method}
\alias{[,VirtualPeriodicAutocovarianceModel,missing,numeric,ANY-method}
\alias{[,VirtualPeriodicAutocovarianceModel,numeric,missing,ANY-method}
\alias{[,VirtualPeriodicAutocovarianceModel,numeric,numeric,ANY-method}
\alias{[,PeriodicTS,missing,missing,ANY-method}
\alias{[,PeriodicMTS,ANY,ANY,ANY-method}
\alias{[,PeriodicMTS,AnyDateTime,ANY,ANY-method}
\alias{[,PeriodicMTS,AnyDateTime,missing,ANY-method}
\alias{[,PeriodicTS,AnyDateTime,missing,ANY-method}

\title{Indexing of objects from classes in package pcts}
\description{Indexing of objects from classes in package pcts.}
\section{Methods}{
  \describe{
    \item{\code{signature(x = "BasicCycle", i = "ANY", j = "missing", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "BasicCycle", i = "missing", j = "missing", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicMTS", i = "ANY", j = "missing", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicMTS", i = "missing", j = "missing", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicVector", i = "ANY", j = "ANY", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicVector", i = "ANY", j = "missing", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicVector", i = "missing", j = "ANY", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicVector", i = "missing", j = "missing", 
          drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualPeriodicAutocovarianceModel", i = "missing", 
          j = "missing", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualPeriodicAutocovarianceModel", i = "missing", 
          j = "numeric", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualPeriodicAutocovarianceModel", i = "numeric", 
          j = "missing", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualPeriodicAutocovarianceModel", i = "numeric", 
          j = "numeric", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicTS", i = "missing", j = "missing", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}
\item{\code{signature(x = "PeriodicMTS", i = "ANY", j = "ANY", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicMTS", i = "AnyDateTime", j = "ANY", drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicMTS", i = "AnyDateTime", j = "missing", 
          drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicTS", i = "AnyDateTime", j = "missing", 
          drop = "ANY")}}{{
%%  ~~describe this method here~~
}
}}}
\author{Georgi N. Boshnakov}
\keyword{methods}
