\name{[<--methods}
\docType{methods}
\alias{[<--methods}
\alias{[<-,pc.armaPQ,ANY,ANY,ANY-method}
%\alias{[<-,pc.Model.WeaklyStat,ANY,ANY,ANY-method}
\alias{[<-,slMatrix,ANY,ANY,ANY-method}
\alias{[<-,ANY,ANY,ANY,ANY-method}
%\alias{[<-,PeriodicAutocovarianceModel,ANY,ANY,ANY-method}
\alias{[<-,PeriodicVector,ANY,ANY,ANY-method}
\alias{[<-,PeriodicVector,missing,ANY,ANY-method}
\alias{[<-,BasicCycle,ANY,missing,ANY-method}
\alias{[<-,BasicCycle,missing,missing,ANY-method}
\alias{[<-,PeriodicAutocovarianceModel,ANY,ANY,ANY-method}
\title{Index assignments for objects from classes in package pcts}
\description{Index assignments for objects from classes in package pcts.}
\section{Methods}{
%%% \item{\code{signature(x = "pc.Model.WeaklyStat", i = "ANY", j = "ANY", value = "ANY")}}{
%%% %%  ~~describe this method here~~
%%% }

%%% \item{\code{signature(x = "slMatrix", i = "ANY", j = "ANY", value = "ANY")}}{
%%% 
%%%   See % \code{\link{[-methods}} and
%%%   \code{\link{slMatrix-class}}.
%%% }
\describe{

\item{\code{signature(x = "PeriodicVector", i = "ANY", j = "ANY", value = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicVector", i = "missing", j = "ANY", value = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "BasicCycle", i = "ANY", j = "missing", value = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "BasicCycle", i = "missing", j = "missing", value = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "PeriodicAutocovarianceModel", i = "ANY", j = "ANY",
          value = "ANY")}}{{
%%  ~~describe this method here~~
}
}}}
\author{Georgi N. Boshnakov}
\keyword{methods}
