\name{[[-methods}
\docType{methods}
\alias{[[-methods}
\alias{[[,PeriodicMTS,ANY-method}
\alias{[[,PeriodicAutocovarianceModel,numeric-method}
\alias{[[,PeriodicMTS,ANY,ANY-method}
\alias{[[,VirtualPeriodicAutocovarianceModel,numeric,ANY-method}
\title{Methods for function\code{`[[`} in package 'pcts'}
\description{Methods for function\code{`[[`} in package 'pcts'.}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicMTS", i = "ANY", j = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualPeriodicAutocovarianceModel", i = "numeric", 
          j = "ANY")}}{{
%%  ~~describe this method here~~
}
}}}
\keyword{methods}
