\name{$-methods}
\docType{methods}
\alias{$-methods}
\alias{$,PeriodicMTS-method}
\title{Methods for function\code{$} in package 'pcts'}
\description{Methods for function\code{$} in package 'pcts'.}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicMTS")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
