\name{as_date-methods}
\docType{methods}
\alias{as_date-methods}
\alias{as_date,PeriodicTimeSeries-method}
\alias{as_date,ANY-method}
\alias{as_date,character-method}
\alias{as_date,Cyclic-method}
\alias{as_date,numeric-method}
\alias{as_date,POSIXt-method}
\title{Replace methods for as_date in package pcts}
\description{Replace methods for as_date in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicTimeSeries")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(x = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "character")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "Cyclic")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "numeric")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "POSIXt")}}{{
%%  ~~describe this method here~~
}
}}}
\keyword{methods}
