\name{backwardPartialCoefficients-methods}
\docType{methods}
\alias{backwardPartialCoefficients-methods}
\alias{backwardPartialCoefficients,VirtualPeriodicAutocovariances-method}
\title{
  Compute periodic backward partial coefficients
}
\description{
  Methods for computation of periodic backward partial coefficients.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "VirtualPeriodicAutocovariances")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
