\name{backwardPartialVariances-methods}
\docType{methods}
\alias{backwardPartialVariances-methods}
\alias{backwardPartialVariances,VirtualPeriodicAutocovariances-method}
\title{
  Compute periodic backward partial variances
}
\description{
  Compute periodic backward partial variances.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "VirtualPeriodicAutocovariances")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
