\name{partialAutocorrelations-methods}
\docType{methods}
\alias{partialAutocorrelations}
\alias{partialAutocorrelations-methods}
\alias{partialAutocorrelations,PeriodicAutocovariances,ANY,missing-method}
\alias{partialAutocorrelations,VirtualPeriodicAutocovariances,ANY,ANY-method}

\title{
  Compute periodic partial autocorrelations
}
\description{
  Methods for computation of periodic partial autocorrelations.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicAutocovariances", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "VirtualPeriodicAutocovariances", maxlag = "ANY", lag_0 = "ANY")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link[sarima]{partialAutocorrelations}} in package \pkg{sarima}
  for further details.
  
  \code{\link{partialVariances}},
  \code{\link{partialAutocovariances}} 
}
\keyword{methods}
