\name{partialCoefficients-methods}
\docType{methods}
\alias{partialCoefficients}
\alias{partialCoefficients-methods}
\alias{partialCoefficients,PeriodicArModel-method}
\alias{partialCoefficients,VirtualPeriodicAutocovariances-method}

\title{
  Compute periodic partial coefficients
}
\description{
  Methods for computation of periodic partial coefficients.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "PeriodicArModel")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "VirtualPeriodicAutocovariances")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
