\name{partialVariances-methods}
\docType{methods}
\alias{partialVariances}
\alias{partialVariances-methods}
\alias{partialVariances,VirtualPeriodicAutocovariances-method}
\title{
  Compute periodic partial variances
}
\description{
  Compute periodic partial variances.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "VirtualPeriodicAutocovariances")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link[sarima]{partialVariances}} in package \pkg{sarima}
  for further details.
  
  \code{\link{partialAutocorrelations}} for for partial autocorrelations.
}
\keyword{methods}
