\name{pcts-deprecated}
\alias{pcts-deprecated}
\alias{mCpar}
\alias{sim_arAcf}
\alias{ptildeorders}

\title{Deprecated Functions in Package \pkg{pcts}}
\description{

  These functions are marked for removal and are provided temporarily
  for compatibility with older versions of package \pkg{pcts} only. Use
  the recommended renamed or new functions instead.

}
%\usage{
%}
%\arguments{
%  \item{...}{arguments pass on to the new function.}
%}
\details{
  
  \describe{

    \item{mCpar}{has been renamed to \code{sim_parCoef}}
    \item{sim_arAcf}{has been renamed to \code{sim_parAcvf}}

  }
}
