% see Repeated descriptions of re-exported functions #408 at
%   https://github.com/klutometis/roxygen/issues/408
\docType{import}
\name{pcts_reexports}
\alias{pcts_reexports}
\alias{slMatrix}

\title{Objects exported from other packages}
\description{
  These objects are imported from other packages. Follow the links
  below to see their documentation.

  \describe{
    \item{lagged}{\code{\link[lagged]{slMatrix}}}
  }
}
\keyword{internal}
