\name{permodelmf}
\alias{permodelmf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Compute the multi-companion form of a per model}
\description{Compute the multi-companion form of a per model.}
\usage{
permodelmf(permodel, update = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{permodel}{a model.}
  \item{update}{If TRUE store the multi-companion form in
    \code{permodel} and return the whole model, otherwise simply return
    the multi-companion form.}
}
\details{todo:}
\value{
  the multi-companion form of the model or the updated model, as
  described in Details.
}
\references{
  \insertRef{boshnakov2009gen}{pcts}
}
\author{Georgi N. Boshnakov}
% \note{
% %%  ~~further notes~~
% }
%
% %% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ permodel }
