\name{unitCycle<--methods}
\docType{methods}
\alias{unitCycle<--methods}
\alias{unitCycle<-,Cyclic-method}
\alias{unitCycle<-,SimpleCycle-method}

\title{Methods for \code{"unitCycle<-"()} in package pcts}
\description{Methods for \code{"unitCycle<-"()} in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "Cyclic")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "SimpleCycle")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
