\name{unitSeason<--methods}
\docType{methods}
\alias{unitSeason<--methods}
\alias{unitSeason<-,Cyclic-method}
\alias{unitSeason<-,SimpleCycle-method}
\title{Methods for \code{"unitSeason<-"()} in package pcts}
\description{Methods for \code{"unitSeason<-"()} in package pcts.}
\section{Methods}{
\describe{

\item{\code{signature(x = "Cyclic")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "SimpleCycle")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
